Moved InputContext to parent folder
